% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aal.R
\docType{data}
\name{AAL_3mm}
\alias{AAL_3mm}
\title{Automated Anatomical Labeling in 3mm}
\format{A BrcParcellation object of size 61 x 73 x 61 voxels, of which
there are 47,636 voxels in representing the brain (17.5%). The parcellation is
not symmetric.}
\source{
TO BE FILLED
}
\description{
Automated Anatomical Labeling in 3mm
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}

