% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aal.R
\docType{data}
\name{AAL_tab}
\alias{AAL_tab}
\title{Automated Anatomical Labeling Table}
\format{A table with 116 rows and 3 variables
\describe{
  \item{Index}{the value used in brcdata to encode the partitions for AAL}
  \item{Short.Name}{short name of the scientific name for the partition}
  \item{Full.Name}{full name of the scientific name for the partition. Only the first
  90 regions are recorded.}
}}
\source{
TO BE FILLED
}
\description{
Automated Anatomical Labeling Table
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}

