% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_parcellation.R
\name{apply_parcellation}
\alias{apply_parcellation}
\title{Apply parcellation}
\usage{
apply_parcellation(arr, parcel_arr)
}
\arguments{
\item{arr}{functional data of subject}

\item{parcel_arr}{parcellation}
}
\value{
a matrix with number of rows equal to number of time slices and number
or columns equal to number of parcellations
}
\description{
Apply parcellation
}

