% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_graph.R
\name{estimate_graph}
\alias{estimate_graph}
\title{Estimate Graph}
\usage{
estimate_graph(mat, heuristic = graph_heuristic_percentage, ...)
}
\arguments{
\item{mat}{Input matrix}

\item{heuristic}{Heuristic to do model selection}

\item{...}{Additional parameters for heuristic}
}
\value{
Graph encoded as sparse matrix
}
\description{
Estimate Graph
}

