% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_difference.R
\name{graph_difference}
\alias{graph_difference}
\title{Graph difference}
\usage{
graph_difference(mat1, mat2)
}
\arguments{
\item{mat1}{graph encoded as sparse matrix}

\item{mat2}{graph encoded as sparse matrix}
}
\value{
The percentage difference and list of edges in one graph but not the other
}
\description{
Graph difference
}

