% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_graph.R
\name{graph_heuristic_percentage}
\alias{graph_heuristic_percentage}
\title{Graph Heurisitic}
\usage{
graph_heuristic_percentage(vec, percentage = 0.1)
}
\arguments{
\item{vec}{vector of sparsity values}

\item{percentage}{Target sparsity percentage}
}
\value{
Index of the mat_list
}
\description{
Graph Heurisitic
}

