% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_graph.R
\name{median_graph}
\alias{median_graph}
\title{Median graph}
\usage{
median_graph(mat_list, heuristic = median_heuristic, ...)
}
\arguments{
\item{mat_list}{list of graphs encoded as sparse matrices}

\item{heuristic}{heuristic to choose rank of value for median graph}

\item{...}{additional parameters for median graph}
}
\value{
a graph encoded as sparse matrix
}
\description{
Median graph
}

