% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_graph.R
\name{median_heuristic}
\alias{median_heuristic}
\title{Median heuristic}
\usage{
median_heuristic(sum_mat, percentage = 0.9)
}
\arguments{
\item{sum_mat}{The matrix that with values equal to number of instances per edge}

\item{percentage}{threshold for how large the quantile is}
}
\value{
a value of number of instances to threshold below
}
\description{
Median heuristic
}

