% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aal.R
\docType{data}
\name{AAL_3mm}
\alias{AAL_3mm}
\title{Automated Anatomical Labeling in 3mm}
\format{A BrcParcellation object of size 61 x 73 x 61 voxels, of which
there are 47,636 voxels in representing the brain (17.5%). The parcellation is
not symmetric.
\describe{
  \item{dim3d}{A 3-element vector for the 3D dimensions}
  \item{partition}{A vector of length prod(dim3d) having values 0,...,n
  where n is the total number of parcellations}
}}
\source{
TO BE FILLED
}
\description{
Automated Anatomical Labeling in 3mm
}
\author{
Kevin Lin \email{kevinl1@andrew.cmu.edu}
}
\keyword{data}

