% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_graph.R
\name{graph_heuristic_percentage}
\alias{graph_heuristic_percentage}
\title{Graph Heurisitic}
\usage{
graph_heuristic_percentage(mat_list, percentage = 0.4)
}
\arguments{
\item{mat_list}{List of graphs encoded as sparse matrices}

\item{percentage}{Target sparsity percentage}
}
\value{
Index of the mat_list
}
\description{
Graph Heurisitic
}

